Added autodetection if file is updated
If the thired line of the settings.txt file is set to "True" or "true",
etc. the program will ignore the set delay, but update every time it
notices a change in the image file size (what are the odds of 2 large
photos having exactly the same number of bytes?).  At the moment it is
set to look for changes in file size every 4 seconds.